fix: eliminate layout shift by server-side rendering OAuth providers on login page
OAuth providers were being fetched client-side with useEffect, causing layout shifts and poor Core Web Vitals scores. This refactors the login page to use a Server Component wrapper that fetches providers at build time and passes them as props to the client component.

Changes:
- Created auth-providers.ts utility to extract OAuth providers from NextAuth config
- Converted login page.tsx to Server Component wrapper
- Moved client logic to login-client.tsx component
- Removed client-side provider fetching to prevent layout shifts
